Moved contents of mita to ocs module
