+++ pytest_unittests.ini
@@ -1,3 +1,3 @@
 [pytest]
 log_format = %(asctime)s - %(threadName)s - %(name)s - %(levelname)s - %(message)s
-testpaths = mita ocs ocsci resources utility
+testpaths = ocs ocsci resources utility
