+++ tox.ini
@@ -12,7 +12,6 @@ setenv =
 commands = py.test \
     --ignore=tests \
     -c pytest_unittests.ini \
-    --cov=mita \
     --cov=ocs \
     --cov=ocsci \
     --cov=resources \
